Quick fixes

+++ personnel-tracking-entity/Personnel.cs
@@ -15,7 +15,7 @@ namespace personnel_tracking_entity
 
         public int PersonnelId { get; set; }
         public int CompanyId { get; set; }
-        public long PersonnelTypeId { get; set; }
+        public int PersonnelTypeId { get; set; }
         public long IdentityNumber { get; set; }
         public string PersonnelName { get; set; }
         public string PersonnelSurname { get; set; }
